Reorganize files; get rid of the Result#match DSL.
